requirements.txt
python-dotenv
mongox
pyrogram
pymongo[srv]
